+++ packages/voorbeeld-design-tokens/documentation/avatar/ams-avatar.stories.tsx
@@ -1,0 +1,26 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { Avatar } from '@amsterdam/design-system-react';
+
+const meta = {
+  id: 'ams-avatar',
+  title: 'Components/Avatar/Amsterdam',
+  component: Avatar,
+  parameters: { actions: { disable: true } },
+  args: {
+    label: 'Avatar',
+  },
+} satisfies Meta<typeof Avatar>;
+
+type Story = StoryObj<typeof meta>;
+
+export default meta;
+
+export const VoorbeeldTheme: Story = {
+  name: 'Voorbeeld theme',
+  parameters: { theme: 'voorbeeld-theme' },
+};
+
+export const AmsterdamTheme: Story = {
+  name: 'Amsterdam theme',
+  parameters: { theme: 'ams-theme' },
+};